requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
pydantic-settings
email-validator
sqlalchemy
psycopg[binary]
alembic
passlib[bcrypt]
python-dotenv
PyJWT
anyio
attrs
certifi
cffi
chardet
click
dnspython
greenlet
h11
idna
importlib-metadata
iniconfig
Mako
MarkupSafe
numpy
packaging
pandas
pluggy
py
pycparser
pyparsing
python-dateutil
python-editor
python-multipart
pytz
requests
six
sniffio
starlette
text-unidecode
toml
typing_extensions
urllib3
zipp
pytest
Faker
fastapi_sqlalchemy
psycopg2-binary